Moved LiveScore feature to a new package
