+++ Java-Space-Invaders-master/src/com/zetcode/LiveScore/ILiveScoreObserver.java
@@ -1,4 +1,4 @@
-package com.zetcode;
+package com.zetcode.LiveScore;
 
 public interface ILiveScoreObserver {
     public void update(int score);
