+++ Java-Space-Invaders-master/src/com/zetcode/LiveScore/ILiveScoreSubject.java
@@ -1,4 +1,4 @@
-package com.zetcode;
+package com.zetcode.LiveScore;
 
 import java.awt.Graphics;
